+++ apps/zhiji-desktop/src/main.ts
@@ -31,8 +31,8 @@ const createWindow = () => {
 // This method will be called when Electron has finished
 // initialization and is ready to create browser windows.
 // Some APIs can only be used after this event occurs.
-app.on('ready', () => {
-  bootstrap();
+app.on('ready', async () => {
+  await bootstrap();
   createWindow();
 });